Avoid exposing new PopParameters class, change API to match design doc
